Put code dealing with internals in its own file
