+++ src/unwrap_ssl.cpp
@@ -1,0 +1,30 @@
+#include "unwrap_ssl.h"
+
+#define NODE_WANT_INTERNALS 1
+#include <tls_wrap.h>
+#include <js_native_api_v8.h>
+
+using v8::Local;
+using v8::String;
+using v8::Object;
+using v8::Isolate;
+
+class TLSWrap2 : public node::TLSWrap {
+    public: SSL* get_ssl(){ return ssl_.get(); }
+};
+
+std::string v8_local_obj_ctor(Local<Object> obj) {
+    Local<String> ctor = obj->GetConstructorName();
+    return *String::Utf8Value(Isolate::GetCurrent(), ctor);
+}
+
+SSL* unwrap_ssl(napi_env env, Napi::Object socket) {
+    Napi::Value wrap_js = socket.Get("_handle");
+    if (!wrap_js.IsObject())
+        throw Napi::TypeError::New(env, "'_handle' property is not an object");
+
+    Local<Object> wrap_v8 = v8impl::V8LocalValueFromJsValue(wrap_js).As<Object>();
+    if (v8_local_obj_ctor(wrap_v8)!="TLSWrap")
+        throw Napi::TypeError::New(env, "'_handle' property is not TLSWrap");
+    return TLSWrap2::FromJSObject<TLSWrap2>(wrap_v8)->get_ssl();
+}
